Added GET mapping for finding all bike orders and changed bike getter for price to int.

+++ src/main/java/com/example/demo/model/Bike.java
@@ -81,7 +81,7 @@ public class Bike {
         this.frameSize = frameSize;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
